move CurrentView to its own class
