+++ src/main/scala/bifrost/api/http/ApiRouteWithView.scala
@@ -4,8 +4,8 @@ import akka.actor.ActorRef
 import akka.pattern.ask
 import bifrost.history.History
 import bifrost.mempool.MemPool
+import bifrost.nodeView.CurrentView
 import bifrost.nodeView.GenericNodeViewHolder.ReceivableMessages.GetDataFromCurrentView
-import bifrost.nodeView.GenericNodeViewHolder.CurrentView
 import bifrost.state.State
 import bifrost.wallet.Wallet
