+++ src/main/scala/bifrost/nodeView/CurrentView.scala
@@ -1,0 +1,3 @@
+package bifrost.nodeView
+
+case class CurrentView[HIS, MS, VL, MP](history: HIS, state: MS, vault: VL, pool: MP)
